refactor main.go to support dynamic multi-target exporters
